finished tokenizer

+++ src/assembler/lexer/token.rs
@@ -1,8 +1,9 @@
 use crate::instruction::Opcode;
 
 #[allow(dead_code)]
+#[derive(Debug, PartialEq)]
 pub enum Token {
     Op { code: Opcode },
     Register { reg_number: u8 },
-    IntegerOp { value: u16 },
+    IntegerOp { value: i32 },
 }
